Re merge complete - moved to double value for NPeople and sim still works

+++ validated/src/test/java/populationStatistics/dataDistributionsTables/selfCorrecting/SelfCorrectionTests.java
@@ -244,7 +244,7 @@ public class SelfCorrectionTests {
         int count = calcUnfetteredExpectedCount(applied, corrective, targetRate);
 
         if(count > corrective.getForNPeople()) {
-            count = corrective.getForNPeople();
+            count = (int) corrective.getForNPeople();
         }
 
         if(count < 0) {
